### 2024-09-07 #7198 @editorialbot
[PRE REVIEW]: cogsworth: A Gala of COSMIC proportions combining binary stellar evolution and galactic dynamics
**Submitting author:** @TomWagg (<a href="http://orcid.org/0000-0001-6147-5761">Thomas Wagg</a>)
**Repository:** https://github.com/TomWagg/cogsworth
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0.0
