### 2024-09-07 #7198 @dfm
In the meantime, it looks like there are some paper build issues. My guess is that it is somehow discovering the ApJS paper instead of the JOSS one. Perhaps you can try renaming the ApJS tex file to something else (and maybe clean up some of the extra temporary files in that directory while you're at it!)?

I also want to note that this submission is made in parallel with a AAS publication. We need to remember to make reviewers aware of our [JOSS collaboration with AAS publishing](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing), and that as part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS.
