+++ paper/ApJS/apjs_paper.tex
@@ -2,7 +2,7 @@
 \PassOptionsToPackage{oneside}{geometry}
 \PassOptionsToPackage{pdfencoding=auto}{hyperref}
 \PassOptionsToPackage{mathlines}{lineno}
-\documentclass[twocolumn, twocolappendix, oneside]{aastex631}
+\documentclass[twocolumn, twocolappendix, oneside, linenumbers]{aastex631}
 \received{\today}
 \shorttitle{\cogsworth}
 \shortauthors{Wagg, Breivik, Renzo \& Price-Whelan}
@@ -118,16 +118,6 @@ Single massive stars are not expected to migrate far from their birth location b
 
 The use of positions and kinematics as tracers of binary evolution has been considered in the past. Recent work has shown the importance of accounting for the galactic potential, which can change the velocity of kicked objects \citep[e.g.,][]{Disberg+2024:2024A&A...687A.272D}. It is also important to consider the inclination or timing of a supernova kick relative to the galactic orbit, since, for example, a kick out of the galactic plane at an object's highest galactic vertical position will have a strong effect on its final position. Failing to consider impacts from both a galactic potential and kicks (i.e. velocity impulses) will lead to misleading conclusions regarding the final spatial distributions of the population. Some studies have considered using the Galactic potential at the present-day position of objects to place a lower limit on the peculiar velocity at birth and constrain supernova kicks \citep{Repetto+2012:2012MNRAS.425.2799R, Repetto+2015:2015MNRAS.453.3341R, Repetto+2017:2017MNRAS.467..298R, atri:19}, but the accuracy of this method is debated \citep{Mandel+2016:2016MNRAS.456..578M}. Other work has considered the impact of the Galactic potential for individual special cases, rather than at a population level. For example, \citet{Evans+2020:2020MNRAS.497.5344E} considered the orbits of hyper-runaway candidates evolving through the Milky Way potential, whilst \citet{Neuhauser+2020:2020MNRAS.498..899N} developed software for tracing the motion of stars to investigate the recent nearby supernovae that ejected $\zeta$ Ophiuchi.
 
-% We demonstrate the need for accounting for binary interactions and a galactic potential simultaneously in Figure~\ref{fig:zgrid}. We use \cogsworth to simulate a population of massive stars formed in the last 100 Myr in the Milky Way and compare the present day positions of single stars, binary stars and binary stars neglecting the galactic potential. This figure not only demonstrates that binary interactions can significantly alter the spatial distribution of massive stars (comparing the top right quadrant to the left half), but also highlights that neglecting the galactic potential could lead to an overestimate of the impact of binary interactions (comparing the top right and bottom right quadrants).
-
-% \begin{figure}[t]
-%     \addDocsIcon{fig:zgrid}{https://cogsworth.readthedocs.io/en/latest/case_studies/binaries_and_potentials.html}
-%     \centering
-%     \includegraphics[width=\columnwidth]{figures/bin_pot_effects.pdf}
-%     \caption{Binary evolution and the presence of a galactic potential both have a significant impact on the spatial distribution of stars. Each panel shows the spatial distributions of massive stars at present day from a \cogsworth population of single stars (left), binaries (right, top) and binaries evolved without a galactic potential (right, bottom). Density distributions are computed using kernel density estimators (KDE) and shown (in linearly spaced isolevels) up to the 98th percentiles, with the remaining stars plotted as scatter points. Marginal KDEs are shown on the right with a log scale.}
-%     \label{fig:zgrid}
-% \end{figure}
-
[…]
