### 2024-09-07 #7198 @TomWagg
Thanks @dfm - sounds good to me! Ah yeah sorry that folder was a bit of a mess, I've cleaned it up a bit now. (I promise I tested the compilation with the GitHub action and it worked there 😅)
