+++ README.md
@@ -1,12 +1,24 @@
-
-<h1 align='center'>cogsworth</h1>
-<p align="center">cogsworth is a Python package for performing self-consistent population synthesis and galactic dynamics simulations</p>
-<p align="center">
-  <a href='https://cogsworth.readthedocs.io/en/latest/?badge=latest'>
-    <img src='https://readthedocs.org/projects/cogsworth/badge/?version=latest' alt='Documentation Status' />
-</a>
-</p>
-
 <p align="center">  
   <img width=500 src="docs/_static/cogsworth_invite.png">
 </p>
+<h1 align='center'>
+  cogsworth
+  <br>
+  <a href="https://raw.githubusercontent.com/TomWagg/cogsworth/paper/paper/ApJS/apjs_paper.pdf">
+      <img src="https://img.shields.io/badge/read-paper-blue.svg?style=flat" alt="Read the article"/>
+  </a>
+
+  <a href='https://cogsworth.readthedocs.io/en/latest/?badge=latest'>
+      <img src='https://readthedocs.org/projects/cogsworth/badge/?version=latest' alt='Documentation Status' />
+  </a>
+  <a href="mailto:tomjwagg@gmail.com">
+      <img src="https://img.shields.io/badge/contact-authors-blueviolet.svg?style=flat" alt="Email the authors"/>
+  </a>
+</h1>
+<h4 align="center">cogsworth is a Python package for performing seamlessly self-consistent population synthesis and galactic dynamics simulations!</h4>
+
+### Installation
+A quick and easy `pip install cogsworth` will cover most use cases, but we recommend using a conda environment as well. Check out the full instructions [here](https://cogsworth.readthedocs.io/en/latest/pages/install.html).
+
+### Documentation
+You can learn all about cogsworth from our collection of tutorials, examples and in-depth case studies on our online documentation [here](https://cogsworth.readthedocs.io/en/latest/index.html).
