+++ paper/ApJS/apjs_paper.tex
@@ -2,7 +2,7 @@
 \PassOptionsToPackage{oneside}{geometry}
 \PassOptionsToPackage{pdfencoding=auto}{hyperref}
 \PassOptionsToPackage{mathlines}{lineno}
-\documentclass[twocolumn, twocolappendix, oneside, linenumbers]{aastex631}
+\documentclass[twocolumn, twocolappendix, oneside]{aastex631}
 \received{\today}
 \shorttitle{\cogsworth}
 \shortauthors{Wagg, Breivik, Renzo \& Price-Whelan}
