### 2024-09-19 TomWagg/cogsworth#137 @Her00000
Error loading snapshot
`snap = cogsworth.hydro.utils.prepare_snapshot("/Users/zzy/snapshot_600.hdf5")`

```
{
	"name": "ValueError",
	"message": "Insufficient particles around center to get velocity",
	"stack": "---------------------------------------------------------------------------
RuntimeError                              Traceback (most recent call last)
File ~/miniconda3/envs/cogsworth/lib/python3.10/site-packages/cogsworth/hydro/utils.py:47, in prepare_snapshot(path, halo_params)
     46 try:
---> 47     h = snap.halos()
     48     pynbody.analysis.halo.center(h[1], **halo_params)

File ~/miniconda3/envs/cogsworth/lib/python3.10/site-packages/pynbody/snapshot/simsnap.py:771, in SimSnap.halos(self, *args, **kwargs)
    769         pass
--> 771 raise RuntimeError(\"No halo catalogue found for %r\" % str(self))

RuntimeError: No halo catalogue found for '<SimSnap \"/Users/zzy/snapshot_600.hdf5\" len=35940848>'

Cell In[16], line 1
----> 1 snap = cogsworth.hydro.utils.prepare_snapshot(\"/Users/zzy/snapshot_600.hdf5\")

File ~/miniconda3/envs/cogsworth/lib/python3.10/site-packages/cogsworth/hydro/utils.py:55, in prepare_snapshot(path, halo_params)
     53     with warnings.catch_warnings():
     54         warnings.filterwarnings(\"ignore\", message=\".*units.*\")
---> 55         pynbody.analysis.halo.center(snap, **halo_params)
     57 # orient the snapshot face-on and side-on
     58 side_on = pynbody.analysis.angmom.sideon(snap, cen=(0, 0, 0))

File ~/miniconda3/envs/cogsworth/lib/python3.10/site-packages/pynbody/analysis/halo.py:336, in center(sim, mode, retcen, vel, cen_size, move_all, wrap, **kwargs)
    333     tx = transformation.inverse_translate(target, cen)
    335 if vel:
--> 336     velc = vel_center(sim, cen_size=cen_size, retcen=True)
    337     tx = transformation.inverse_v_translate(tx, velc)
    339 return tx

File ~/miniconda3/envs/cogsworth/lib/python3.10/site-packages/pynbody/analysis/halo.py:252, in vel_center(sim, mode, cen_size, retcen, move_all, **kwargs)
    249     cen = sim.gas[filt.Sphere(cen_size)]
    250 if len(cen) < 5:
    251     # very weird snapshot, or mis-centering!
--> 252     raise ValueError(\"Insufficient particles around center to get velocity\")
    254 vcen = (cen['vel'].transpose() * cen['mass']).sum(axis=1) / \\
    255     cen['mass'].sum()
    256 vcen.units = cen['vel'].units

ValueError: Insufficient particles around center to get velocity"
}

```

The snapshot from https://users.flatironinstitute.org/~chayward/fire2_public_release/core/m11h_res7100/output/, should be as same as in the example, but will go wrong.
