### 2024-09-19 TomWagg/cogsworth#137 @TomWagg
Hi @Her00000, thanks for the issue, I think I know what the problem is, it comes from the configuration in pynbody that I forgot to mention in the docs sorry.

Could you please try creating a file called `.pynbodyrc` in your home directory with the contents of
```
[gadget-units]
pos: kpc a h^-1

[gadgethdf-type-mapping]
dm: PartType1
```
And re-run. Let me know if that fixes it!
