### 2024-09-23 54b429a
always use initC if it's available
