### 2024-10-23 #7400 @editorialbot
[REVIEW]: cogsworth: A Gala of COSMIC proportions combining binary stellar evolution and galactic dynamics
**Submitting author:** @TomWagg (<a href="http://orcid.org/0000-0001-6147-5761">Thomas Wagg</a>)
**Repository:** https://github.com/TomWagg/cogsworth
**Branch with paper.md** (empty if default branch): paper
**Version:** 2.0.1
