### 2024-10-25 #7400 @augustfly
The AAS ApJS article DOI will be "10.3847/1538-4365/ad8b1f"
