### 2024-11-04 TomWagg/cogsworth#137 @nicolagaspari
_This issue is flagged as part of the JOSS review [#7400](https://github.com/openjournals/joss-reviews/issues/7400)_

I had the same problem and your solution solved it, so worth adding it to the docs.
