### 2024-11-04 TomWagg/cogsworth#137 @TomWagg
Thanks for the reminder, I'll add this. I may also just add an error catch for this that points to the docs so it makes the error less confusing.
