### 2024-11-05 205e569
update ce-potential case study
