### 2024-11-05 709b175
update binaries-pot case study
