+++ docs/pages/case_studies.rst
@@ -4,8 +4,7 @@ Case Studies
 
 This section contains a collection of case studies that demonstrate some more detailed uses of ``cogsworth``.
 These are designed to be more in-depth than the tutorials and show how to use the package in a more complex way.
-Note that they also will take longer to run if you want to reproduce the results (on the order of 15 minutes each)
-and some require downloading some additional data.
+Note that they also will take a lot longer to run if you want to reproduce the results and some require downloading some additional data.
 
 .. grid:: 1 1 2 2
     :gutter: 3
