+++ paper/ApJS/bibs/paper.bib
@@ -4380,3 +4380,232 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Rossi+2015:2015A&C....12...11R,
+       author = {{Rossi}, L.~J.},
+        title = "{NIGO: A Numerical Integrator of Galactic Orbits}",
+      journal = {Astronomy and Computing},
+     keywords = {Galactic dynamics, Galaxy models},
+         year = 2015,
+        month = sep,
+       volume = {12},
+        pages = {11-20},
+          doi = {10.1016/j.ascom.2015.03.008},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015A&C....12...11R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Chattopadhyay+2020:2020MNRAS.494.1587C,
+       author = {{Chattopadhyay}, Debatri and {Stevenson}, Simon and {Hurley}, Jarrod R. and {Rossi}, Luca J. and {Flynn}, Chris},
+        title = "{Modelling double neutron stars: radio and gravitational waves}",
+      journal = {\mnras},
+     keywords = {gravitational waves, stars: neutron, pulsars: general, Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
+         year = 2020,
+        month = may,
+       volume = {494},
+       number = {2},
+        pages = {1587-1610},
+          doi = {10.1093/mnras/staa756},
+archivePrefix = {arXiv},
+       eprint = {1912.02415},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.494.1587C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Chattopadhyay+2021:2021MNRAS.504.3682C,
+       author = {{Chattopadhyay}, Debatri and {Stevenson}, Simon and {Hurley}, Jarrod R. and {Bailes}, Matthew and {Broekgaarden}, Floor},
+        title = "{Modelling neutron star-black hole binaries: future pulsar surveys and gravitational wave detectors}",
+      journal = {\mnras},
+     keywords = {gravitational waves, pulsars: general, black hole - neutron star mergers, Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
+         year = 2021,
+        month = jul,
+       volume = {504},
+       number = {3},
+        pages = {3682-3710},
+          doi = {10.1093/mnras/stab973},
+archivePrefix = {arXiv},
+       eprint = {2011.13503},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021MNRAS.504.3682C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Song+2024:2024arXiv240611428S,
+       author = {{Song}, Yuzhe and {Stevenson}, Simon and {Chattopadhyay}, Debatri},
+        title = "{Binary population synthesis of the Galactic canonical pulsar population}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2024,
+        month = jun,
+          eid = {arXiv:2406.11428},
+        pages = {arXiv:2406.11428},
+          doi = {10.48550/arXiv.2406.11428},
+archivePrefix = {arXiv},
+       eprint = {2406.11428},
[…]
