+++ paper/ApJS/apjs_paper.tex
@@ -17,6 +17,7 @@
 \usepackage{fontawesome5}
 \usepackage{wrapfig}
 \usepackage[figuresright]{rotating}
+% \usepackage[finalizecache,cachedir=.]{minted}
 \usepackage[frozencache,cachedir=minted]{minted}
 
 % remove indents in footnotes
@@ -61,6 +62,8 @@
 \newcommand{\tutorialIconLink}[1]{\href{#1}{\tutorialIcon}}
 \newcommand{\tutorialLink}[2]{\href{#1}{{\color{codecolour}#2}}}
 
+\newcommand{\invisibleedit}[1]{#1}
+
 \begin{document}
 
 % \darkmode{}
@@ -96,10 +99,10 @@
 \email{tomjwagg@gmail.com}
 
 \begin{abstract}
-    We present \cogsworth, an open-source Python tool for producing self-consistent population synthesis and galactic dynamics simulations.
-    With \cogsworth one can (1) sample a population of binaries and star formation history, (2) perform rapid (binary) stellar evolution, (3) integrate orbits through the galaxy and (4) inspect the full evolutionary history of each star or compact object, as well as their positions and kinematics. We include the functionality for post-processing hydrodynamical zoom-in simulations as a basis for galactic potentials and star formation histories to better account for initial spatial stellar clustering and more complex potentials. Alternatively, several analytic models are available for both the potential and star formation history. \cogsworth can transform the intrinsic simulated population to an observed population through the joint application of dust maps, bolometric correction functions and survey selection functions.
+
+    We present \cogsworth, an open-source Python tool for producing self-consistent population synthesis and galactic dynamics simulations. \cogsworth allows users to (1) sample a population of binaries and star formation history, (2) perform rapid (binary) stellar evolution, (3) integrate orbits through the galaxy, and (4) inspect the full evolutionary history of each star or compact object, along with their positions and kinematics. It supports post-processing hydrodynamical zoom-in simulations for more realistic galactic potentials and star formation histories, accounting for initial spatial stellar clustering and complex potentials. Alternatively, several analytic models are available for galactic potentials and star formation histories. \cogsworth can also transform the intrinsic simulated population into an observed population using dust maps, bolometric correction functions, and survey selection functions.
     
-    We provide a detailed explanation of the functionality of \cogsworth and demonstrate its capabilities through a series of use cases: (1) We predict the spatial distribution of compact objects and runaways in both dwarf and Milky-Way-like galaxies, (2) using a star cluster from a hydrodynamical simulation, we show how supernovae can change the orbits of stars in several ways, and (3) we predict the separation of disrupted binary stellar companions on the sky and create a synthetic \gaia colour-magnitude diagram. We also discuss some current limitations and plans for future developments. We designed \cogsworth and its online documentation to provide a powerful tool for constraining binary evolution, but also a flexible and accessible resource for the entire community.
[…]
