+++ paper/JOSS/paper.bib
@@ -1,505 +1,0 @@
-@ARTICLE{Duchene+2013,
-       author = {{Duch{\^e}ne}, Gaspard and {Kraus}, Adam},
-        title = "{Stellar Multiplicity}",
-      journal = {\araa},
-     keywords = {Astrophysics - Solar and Stellar Astrophysics},
-         year = 2013,
-        month = aug,
-       volume = {51},
-       number = {1},
-        pages = {269-310},
-          doi = {10.1146/annurev-astro-081710-102602},
-archivePrefix = {arXiv},
-       eprint = {1303.3028},
- primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ARA&A..51..269D},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Moe+2017,
-       author = {{Moe}, Maxwell and {Di Stefano}, Rosanne},
-        title = "{Mind Your Ps and Qs: The Interrelation between Period (P) and Mass-ratio (Q) Distributions of Binary Stars}",
-      journal = {\apjs},
-     keywords = {binaries: close, binaries: general, stars: evolution, stars: formation, stars: massive, stars: statistics, Astrophysics - Solar and Stellar Astrophysics},
-         year = 2017,
-        month = jun,
-       volume = {230},
-       number = {2},
-          eid = {15},
-        pages = {15},
-          doi = {10.3847/1538-4365/aa6fb6},
-archivePrefix = {arXiv},
-       eprint = {1606.05347},
- primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2017ApJS..230...15M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@INPROCEEDINGS{Offner+2023,
-       author = {{Offner}, S.~S.~R. and {Moe}, M. and {Kratter}, K.~M. and {Sadavoy}, S.~I. and {Jensen}, E.~L.~N. and {Tobin}, J.~J.},
-        title = "{The Origin and Evolution of Multiple Star Systems}",
-     keywords = {Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies},
-    booktitle = {Protostars and Planets VII},
-         year = 2023,
-       editor = {{Inutsuka}, S. and {Aikawa}, Y. and {Muto}, T. and {Tomida}, K. and {Tamura}, M.},
-       series = {Astronomical Society of the Pacific Conference Series},
-       volume = {534},
-        month = jul,
-        pages = {275},
-          doi = {10.48550/arXiv.2203.10066},
-archivePrefix = {arXiv},
-       eprint = {2203.10066},
- primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ASPC..534..275O},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@ARTICLE{Podsiadlowski+1992,
-       author = {{Podsiadlowski}, Ph. and {Joss}, P.~C. and {Hsu}, J.~J.~L.},
-        title = "{Presupernova Evolution in Massive Interacting Binaries}",
-      journal = {\apj},
-     keywords = {Binary Stars, Massive Stars, Stellar Evolution, Supernovae, Computational Astrophysics, Monte Carlo Method, Stellar Envelopes, Stellar Mass Accretion, Stellar Models, Supernova 1987a, Astrophysics},
-         year = 1992,
-        month = may,
-       volume = {391},
-        pages = {246},
-          doi = {10.1086/171341},
[…]

[…]
