### 2024-11-07 TomWagg/cogsworth#142 @nicolagaspari
Documentation of optional software dependencies
_This issue is submitted as part of the JOSS review [#7400](https://github.com/openjournals/joss-reviews/issues/7400)_

**Which documentation page is causing an issue?**
[https://cogsworth.readthedocs.io/en/latest/pages/install.html](https://cogsworth.readthedocs.io/en/latest/pages/install.html)

**What is the problem?**
Some functionalities, examples, and tutorials will not work with the basic installation `pip install cogsworth`. Although this is hinted, I feel it should be stated more clearly.

**Suggested solution (optional)**
When explaining the installation procedure under **Package installation**, clearly state that several functionalities and tutorials will need `pip install cogsworth[extra]` to run.

Furthermore
1) I would suggest adding a section titled **Python dependencies** that reference the "basic" dependencies (eg. `numpy`, `astropy`, `scipy`, `matplotlib`, `gala`) with a subsection titled **Optional dependencies** that lists all the optional dependencies (eg. `dustmaps`, `gaiaunlimited`, `isochrones`, ...).
2) Specify that "basic" dependencies are installed by default with every installation methods, while optional dependencies need either `pip install cogsworth[extra]` (does `cogsworth[all]` work too?) or manual installation.
