### 2024-11-07 TomWagg/cogsworth#142 @TomWagg
This is a great idea, I should definitely clarify these things, thanks!

(For reference `[all]` includes testing and docs dependencies but `[extra]` should include everything for a non-dev users, I'll put this in the docs!)
