### 2024-11-07 e2a9608
redesign install page to talk better about optional deps
