+++ docs/pages/install.rst
@@ -1,3 +1,5 @@
+.. _install:
+
 *******
 Install
 *******
