### 2024-11-07 306db1d
add warning to the CMD example too
