### 2024-11-07 97c4eed
shell escapes for extras installs (close #144)
