### 2024-11-12 TomWagg/cogsworth#146 @nicolagaspari
Missing descriptions in the docs
_This issue is submitted as part of the JOSS review [#7400](https://github.com/openjournals/joss-reviews/issues/7400)_

**Which documentation page is causing an issue?**
[https://cogsworth.readthedocs.io/en/latest/pages/modules.html](https://cogsworth.readthedocs.io/en/latest/pages/modules.html)

**What is the problem?**
The documentation of a few classes do not provide adequate descriptions for attributes and/or methods (eg. [pop.Population](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.pop.Population.html), [sfh.StarFormationHistory](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.sfh.StarFormationHistory.html)).
This is not a problem for most of them because the names are descriptive enough. However, for eg. [pop.Population.bpp](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.pop.Population.html#cogsworth.pop.Population.bpp) it would be very useful to have a reference at hand for each column and its possible values (especially `kstar` and `evol_type`).

**Suggested solution (optional)**
For the purpose of the review, I recommend improving the documentation of the items below.
1) [utils.translate_COSMIC_tables](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.utils.translate_COSMIC_tables.html#cogsworth.utils.translate_COSMIC_tables) : I suggest having a table that shows numerical types and human-readable types, for both `kstar` and `evol_type`;
2) [classify.list_classes](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.classify.list_classes.html#cogsworth.classify.list_classes) : same as above for classes (short names, long names, and criteria);
3) [pop.Population.bpp](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.pop.Population.html#cogsworth.pop.Population.bpp) : same as above for column names, plus a reference to the lists in [utils.translate_COSMIC_tables](https://cogsworth.readthedocs.io/en/latest/api/cogsworth.utils.translate_COSMIC_tables.html#cogsworth.utils.translate_COSMIC_tables) for `kstar` and `evol_type`.
