### 2024-11-12 #7400 @nicolagaspari
Dear @ivastar, @TomWagg,

Here is my report for this submission.

The review was straightforward and I ticked almost all the checkboxes in one go. The code is easy to read, runs smoothly, and the documentation is quite exhaustive. I was able to follow the documentation for installing and running `cogsworth` as expected, except for a few minor details for which I either raised or flagged issues on the target repository (see below). The tutorials and examples provided are plenty, and thorough, and seem to cover all the functionalities the code provides. The paper is clear and concise although I would improve it on the state of the field (see below).

Overall I'm very positive about `cogsworth`, and at this stage I only have two outstanding issues which are minor and should be easy to fix.

**I) Installation and API documentation**

I had a couple of minor comments on the documentation, for which I opened and flagged issues on the `cogsworth` repository (namely [#137](https://github.com/TomWagg/cogsworth/issues/137), [#142](https://github.com/TomWagg/cogsworth/issues/142), [#146](https://github.com/TomWagg/cogsworth/issues/146)). I think only the last one still needs a fix.

**II) State of the field**

The paper does not list recent works that employed similar methods, nor it provides a comparison with them. I think that showcasing recent works and comparing them to `cogsworth` could really highlight what this code can do and why it is timely.

For instance, recent relevant works are [Sweeney+22](https://arxiv.org/abs/2210.04241) and [Sweeney+24](https://arxiv.org/abs/2403.14612) on Galactic neutron stars and stellar-mass black holes, [Gaspari+24a](https://arxiv.org/abs/2310.14773) and [Disberg+24b](https://arxiv.org/abs/2407.13667) on Galactic binary neutron stars, [Zevin+20](https://arxiv.org/abs/1910.03598) and [Mandhai+22](https://arxiv.org/abs/2109.09714) on gamma-ray bursts progenitors, [Abbott+17](https://arxiv.org/abs/1710.05838) and [Gaspari+24b](https://arxiv.org/abs/2410.19480) on GW170817/GRB170817A. All of these are similar in that they model the galactic potentials, seed synthetic populations in them, then simulate the galactic trajectories and constrain some population of kicked objects. However, each methodology has its own features beside being applied to different objects. `Cogsworth` provides a unified framework that incorporates population synthesis, spectral synthesis, galactic dynamics (potentially for custom star formation histories), can be interfaced with cosmological simulations, and it is general enough that it could find application in all the cases mentioned above. I think that the paper (and the code) could really benefit from a paragraph on this.
