### 2024-11-13 TomWagg/cogsworth#146 @TomWagg
Great point - I've been relying a lot on the COSMIC documentation for this but it definitely makes sense to explain this properly in the docs. I'll add this soon :)
