### 2024-11-13 71b231a
start adding attribute docstrings
