### 2024-11-13 1d5b49b
more docstrings, don't need initial attributes list
