+++ cogsworth/pop.py
@@ -79,47 +79,6 @@ class Population():
         Whether to store the entire orbit for each binary, by default True. If not then only the final
         PhaseSpacePosition will be stored. This cuts down on both memory usage and disk space used if you
         save the Population (as well as how long it takes to reload the data).
-
-    Attributes
-    ----------
-    mass_singles : `float`
-        Total mass in single stars needed to generate population
-    mass_binaries : `float`
-        Total mass in binaries needed to generate population
-    n_singles_req : `int`
-        Number of single stars needed to generate population
-    n_bin_req : `int`
-        Number of binaries needed to generate population
-    bpp : :class:`~pandas.DataFrame`
-        Evolutionary history of each binary
-    bcm : :class:`~pandas.DataFrame`
-        Final state of each binary
-    initC : :class:`~pandas.DataFrame`
-        Initial conditions for each binary
-    kick_info : :class:`~pandas.DataFrame`
-        Information about the kicks that occur for each binary
-    orbits : `list` of :class:`~gala.dynamics.Orbit`
-        The orbits of each system within the galaxy from its birth until :attr:`max_ev_time` with timesteps of
-        :attr:`timestep_size`. This list will have length = `len(self) + self.disrupted.sum()`, where the
-        first section are for bound binaries and disrupted primaries and the last section are for disrupted
-        secondaries
-    classes : `list`
-        The classes associated with each produced binary (see :meth:`~cogsworth.classify.list_classes` for a
-        list of available classes and their meanings)
-    final_pos, final_vel : :class:`~astropy.quantity.Quantity`
-        Final positions and velocities of each system in the galactocentric frame.
-        The first `len(self)` entries of each are for bound binaries or primaries, then the final
-        `self.disrupted.sum()` entries are for disrupted secondaries. Any missing orbits (where orbit=None
-        will be set to `np.inf` for ease of masking.
-    final_bpp : :class:`~pandas.DataFrame`
-        The final state of each binary (taken from the final entry in :attr:`bpp`)
-    disrupted : :class:`~numpy.ndarray` of `bool`
-        A mask on the binaries of whether they were disrupted
-    observables : :class:`~pandas.DataFrame`
-        Observables associated with the final binaries. See `get_photometry` for more details on the columns
-    bin_nums : :class:`~numpy.ndarray`
-        An array containing the unique COSMIC `bin_nums` of each binary in the population - these can be
-        used an indices for the population
     """
     def __init__(self, n_binaries, processes=8, m1_cutoff=0, final_kstar1=list(range(16)),
                  final_kstar2=list(range(16)), sfh_model=sfh.Wagg2022, sfh_params={},
@@ -392,6 +351,11 @@ class Population():
         or :attr:`initC`, or :attr:`initial_binaries` if available. If none of these are available, an
         error is raised.
 
+        Returns
+        -------
+        bin_nums : :class:`~numpy.ndarray`
+            An array of the unique identifiers for each binary in the population.
+
         Raises
         ------
         ValueError
@@ -410,6 +374,21 @@ class Population():
 
     @property
     def initial_galaxy(self):
+        """The initial sampled galactic star formation history.
+
+        This contains the initial conditions for the galaxy that was sampled to generate the population. This
+        includes the birth times, positions, velocities, and metallicities of the stars in the galaxy.
+
+        Returns
+        -------
+        initial_galaxy : :class:`~cogsworth.sfh.StarFormationHistory`
+            The initial galaxy that was sampled to generate the population.
+
+        Raises
+        ------
+        ValueError
+            If no galaxy has been sampled yet.
+        """
         if self._initial_galaxy is None and self._file is not None:
             self._initial_galaxy = sfh.load(self._file, key="initial_galaxy")
             self.sfh_model = self._initial_galaxy.__class__
[…]
