### 2024-11-13 6b081ed
almost all docstrings for attributes in Pop added
