### 2024-11-13 efd603a
document classes better
