### 2024-11-13 7d1181b
missed some functions from __all__ (classic)
