### 2024-11-13 1c28763
add sfh attribute docstrings
