### 2024-11-13 TomWagg/cogsworth#146 @TomWagg
@nicolagaspari FYI I have not added explicit tables for the COSMIC columns but instead linked to the COSMIC docs because the columns are liable to change frequently. Similarly, I've highlighted users can use `cogsworth.classify.list_classes` to get the most up to date list of classes since I add to this a fair bit.

This PR should otherwise fix the things you highlighted were missing :)
