### 2024-11-16 77f8eef
update paper with more state of the field
