+++ paper/JOSS/paper.md
@@ -38,7 +38,7 @@ aas-journal: #Astrophysical Journal <- The name of the AAS journal.
 
 # Summary
 
-We present `cogsworth`, an open-source Python tool for producing self-consistent population synthesis and galactic dynamics simulations. With `cogsworth` one can (1) sample a population of binaries and star formation history, (2) perform rapid (binary) stellar evolution, (3) integrate orbits through the galaxy and (4) inspect the full evolutionary history of each star or compact object, as well as their positions and kinematics. We include the functionality for post-processing hydrodynamical zoom-in simulations as a basis for galactic potentials and star formation histories to better account for initial spatial stellar clustering and more complex potentials. Alternatively, several analytic models are available for both the potential and star formation history. `cogsworth` can transform the intrinsic simulated population to an observed population through the joint application of dust maps, bolometric correction functions and survey selection functions.
+We present `cogsworth`, an open-source Python tool for producing self-consistent population synthesis and galactic dynamics simulations. With `cogsworth` one can (1) sample a population of binaries and star formation history, (2) perform rapid (binary) stellar evolution, (3) integrate orbits through the galaxy and (4) inspect the full evolutionary history of each star or compact object, as well as their positions and kinematics. We include the functionality for post-processing hydrodynamical zoom-in simulations as a basis for galactic potentials and star formation histories to better account for initial spatial stellar clustering and more complex potentials. Alternatively, several analytic models are available for both the potential and star formation history. `cogsworth` can transform the intrinsic simulated population to an observed population through the joint application of dust maps, bolometric correction functions and survey selection functions. This paper is jointly published in ApJS [@cogsworth_full].
 
 # Statement of need
 
@@ -46,9 +46,11 @@ The majority of stars are born in binaries and multiple star systems (e.g., @Duc
 
 Single massive stars are not expected to migrate far from their birth location before reaching core-collapse due to their short lifetimes ($\lesssim50$\,Myr, e.g., @zapartas:17). However, binary stars may disrupt after an initial supernova event, ejecting the secondary star from the system with its orbital velocity (e.g., @Blaauw+1961; @Eldridge+2011; @Renzo+2019). Thus, close massive binaries that disrupt can lead to the displacement of secondary stars significantly farther from star-forming regions. The present-day positions and kinematics of massive stars and binary products are therefore strongly impacted by changes in binary physics that alter the separation prior to supernova. This means that comparing simulations of positions and kinematics of stars and compact objects to observations will enable constraints on binary stellar evolution parameters.
 
[…]
