+++ paper/JOSS/paper.bib
@@ -503,3 +503,237 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.489.3116A},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{underworld,
+       author = {{Sweeney}, David and {Tuthill}, Peter and {Sharma}, Sanjib and {Hirai}, Ryosuke},
+        title = "{The Galactic underworld: the spatial distribution of compact remnants}",
+      journal = {\mnras},
+     keywords = {methods: numerical, astrometry, proper motions, stars: abundances, stars: black holes, stars: neutron, Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2022,
+        month = nov,
+       volume = {516},
+       number = {4},
+        pages = {4971-4979},
+          doi = {10.1093/mnras/stac2092},
+archivePrefix = {arXiv},
+       eprint = {2210.04241},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022MNRAS.516.4971S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Sweeney+2024,
+       author = {{Sweeney}, David and {Tuthill}, Peter and {Krone-Martins}, Alberto and {M{\'e}rand}, Antoine and {Scalzo}, Richard and {Martinod}, Marc-Antoine},
+        title = "{Observing the galactic underworld: predicting photometry and astrometry from compact remnant microlensing events}",
+      journal = {\mnras},
+     keywords = {Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2024,
+        month = jun,
+       volume = {531},
+       number = {2},
+        pages = {2433-2447},
+          doi = {10.1093/mnras/stae1302},
+archivePrefix = {arXiv},
+       eprint = {2403.14612},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024MNRAS.531.2433S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{COMPAS,
+       author = {{Riley}, Jeff and {Agrawal}, Poojan and {Barrett}, Jim W. and {Boyett}, Kristan N.~K. and {Broekgaarden}, Floor S. and {Chattopadhyay}, Debatri and {Gaebel}, Sebastian M. and {Gittins}, Fabian and {Hirai}, Ryosuke and {Howitt}, George and {Justham}, Stephen and {Khandelwal}, Lokesh and {Kummer}, Floris and {Lau}, Mike Y.~M. and {Mandel}, Ilya and {de Mink}, Selma E. and {Neijssel}, Coenraad and {Riley}, Tim and {van Son}, Lieke and {Stevenson}, Simon and {Vigna-G{\'o}mez}, Alejandro and {Vinciguerra}, Serena and {Wagg}, Tom and {Willcox}, Reinhold and {Team Compas}},
+        title = "{Rapid Stellar and Binary Population Synthesis with COMPAS}",
+      journal = {\apjs},
+     keywords = {1622, 154, 1108, 162, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2022,
+        month = feb,
+       volume = {258},
+       number = {2},
[…]
