+++ paper/JOSS/paper.md
@@ -48,7 +48,7 @@ Single massive stars are not expected to migrate far from their birth location b
 
 The use of positions and kinematics as tracers of binary evolution has been considered in the past. Recent work has shown the importance of accounting for the galactic potential, which can change the velocity of kicked objects (e.g. @Disberg+2024). It is also important to consider the inclination or timing of a supernova kick relative to the galactic orbit, since, for example, a kick out of the galactic plane at an object's highest galactic vertical position will have a strong effect on its final position. Failing to consider impacts from both a galactic potential and kicks (i.e. velocity impulses) will lead to misleading conclusions regarding the final spatial distributions of the population. Some studies have considered using the Galactic potential at the present-day position of objects to place a lower limit on the peculiar velocity at birth and constrain supernova kicks [@Repetto+2012; @Repetto+2015; @Repetto+2017; @atri:19], but the accuracy of this method is debated [@Mandel+2016]. Other work has considered the impact of the Galactic potential for individual special cases, rather than at a population level. For example, @Evans+2020 considered the orbits of hyper-runaway candidates evolving through the Milky Way potential, whilst @Neuhauser+2020 developed software for tracing the motion of stars to investigate the recent nearby supernovae that ejected $\zeta$ Ophiuchi. Additionally, @Andrews+2022 considered galactic orbits of synthetic populations to place constraints on black hole natal kicks based on observations of a microlensed black hole.
 
-Additionally, there are several works that consider a full population of objects integrated through a galactic potential. @underworld and @Sweeney+2024 used a combination of `GALAXIA` and `galpy` to predict the spatial distribution of Milky Way black holes and neutron stars. Similarly, several works have combined population synthesis with galactic orbit integration (e.g. using `COMPAS` [@COMPAS] and `NIGO` [@NIGO]) to investigate binary neutron stars and pulsars [@Chattopadhyay+2020; @Chattopadhyay+2021; Gaspari+2024a; @Disberg+2024b; @Song+2024], as well as gamma-ray bursts [@Zevin+2020; Mandai+2022; Gaspari+2024b].
+Additionally, there are several works that consider a full population of objects integrated through a galactic potential. @underworld and @Sweeney+2024 used a combination of `GALAXIA` and `galpy` to predict the spatial distribution of Milky Way black holes and neutron stars. Similarly, several works have combined population synthesis with galactic orbit integration (e.g. using `COMPAS` [@COMPAS] and `NIGO` [@NIGO]) to investigate binary neutron stars and pulsars [@Chattopadhyay+2020; @Chattopadhyay+2021; @Gaspari+2024a; @Disberg+2024b; @Song+2024], as well as gamma-ray bursts [@Zevin+2020; @Mandai+2022; @Gaspari+2024b].
 
 There is a clear need for a unified open-source tool that provides the theoretical infrastructure for making predictions for the positions and kinematics of massive stars and compact objects, placing these systems in the context of their host galaxy and its gravitational potential. `cogsworth` fulfils this need, providing a framework for self-consistent population synthesis and galactic dynamics simulations. The code is applicable to a wide range of binary products, both common and rare, from walkaway and runaway stars to X-ray binaries, as well as gravitational-wave and gamma-ray burst progenitors.
 
