+++ paper/JOSS/paper.bib
@@ -737,3 +737,20 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Gaspari+2024b,
+       author = {{Gaspari}, N. and {Stevance}, H.~F. and {Levan}, A.~J. and {Chrimes}, A.~A. and {Lyman}, J.~D.},
+        title = "{Binary neutron star merger offsets from their host galaxies. GW 170817 as a case study}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2024,
+        month = oct,
+          eid = {arXiv:2410.19480},
+        pages = {arXiv:2410.19480},
+          doi = {10.48550/arXiv.2410.19480},
+archivePrefix = {arXiv},
+       eprint = {2410.19480},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv241019480G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
