+++ paper/JOSS/paper.bib
@@ -663,7 +663,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Gaspari+2024a,
-       author = {{Gaspari}, Nicola and {Levan}, Andrew J. and {Chrimes}, Ashley A. and {Nelemans}, Gijs},
+       author = {{Gaspari}, N. and {Levan}, Andrew J. and {Chrimes}, Ashley A. and {Nelemans}, Gijs},
         title = "{The Galactic neutron star population - II. Systemic velocities and merger locations of binary neutron stars}",
       journal = {\mnras},
      keywords = {binaries: general, gamma-ray burst: general, stars: neutron, Galaxy: stellar content, Galaxy: structure, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, General Relativity and Quantum Cosmology},
