### 2024-11-16 #7400 @TomWagg
Thanks @nicolagaspari - this has been super helpful, I very much appreciate the effort! 🙌🏻

I've gone through and fixed each of the issues you've opened. I've also updated the paper with a new paragraph that covers your suggestions (plus a few extras that are worth highlighting too), great point that this should be included.

Thanks again!
