### 2024-11-17 #7400 @smandhai
Dear @ivastar, @TomWagg ,

My sincere apologies for the delay in getting this review to you. I have now completed my report and here's what I have to say:

As stated with @nicolagaspari, the review process was very straightforward and I ticked the boxes fairly rapidly.

I want to start by saying that `cogsworth`'s documentation is excellent and made the experience smooth to setup and test. I ran through some of the examples and was able to reproduce the reported results.

In the astronomical community, a package such as this has a multitude of uses and has the potential to lead to novel science exploration. Overall, I am very impressed by the accessible computational overhead and the execution time required.

### Functionality

- `cogsworth` installs as expected on Unix systems (including WSL).
- Installation fails on Windows systems (tested on 1 Windows 10 machine and 2 Windows 11). Due to the incompatibility of 'cosmic==3.4.17'. Not instrinsically an issue with this package, but rather its dependency, which prevents complete installation. Issue has been raised in the `cosmic` repo  (here :[https://github.com/COSMIC-PopSynth/COSMIC/issues/682])

### Statement of Need + Statement of Field
The manuscript does a good job of highlighting existing studies exploring existing codes/methods. Its purpose as a generalised open-source tool for simulating stellar kinematics and evolution fills an existing gap in the field. The additional interfacing with cosmological simulations, account for dust attentuation grant it capabilities to simulate mock observations, opening the door to further applications and studies.

Regarding previous comments: Comments raised by @nicolagaspari have been addressed successfully - I echo similar concerns with the original version. These are no longer an issue.

### Final Remarks
I have no concerns besides the raised compatibility with Windows systems. I am otherwise happy to recommend this package and its accompanying documentation for release.
