### 2024-12-05 #7400 @TomWagg
Hi @ivastar - I just wanted to check if you need anything more from me for the paper to be accepted? Thanks! 🙂
