### 2024-12-05 #7400 @ivastar
Hi @TomWagg! I thought you were still addressing issues. The windows installation one is still open. Are you not planning to address it?
