### 2024-12-05 #7400 @TomWagg
Ah my apologies! I had been leaving that to Katie since it was a ``COSMIC`` issue rather than ``cogsworth`` directly. But perhaps I should add a note to the ``cogsworth`` docs in the meantime that explicitly states that we only support Windows when using WSL? I can put that on the installation page now.
