### 2024-12-05 #7400 @ivastar
Well, yes, but it is a required dependency. I assume there is no useful way to run `cogsworth` without `COSMIC` so yes, updating the documentation would be a reasonable solution short of solving the issue and making a pull request to `COSMIC`.
