### 2024-12-05 #7400 @TomWagg
Yes very fair! I don't think the ``COSMIC`` team have any plans to support windows without WSL (since it becomes rather difficult to get the fortran to compile properly and installing WSL isn't too tricky) so updating the docs seems like the best solution. I'll do that right now :)
