### 2024-12-05 #7400 @smandhai
Agreed - the Windows workaround isn't too difficult fortunately. As long as there is a route to installing the package, I think this is fine (personally).
