### 2024-12-05 176ca6f
add windows/WSL install warning
