### 2024-12-05 #7400 @TomWagg
Thanks @smandhai!

I've just merged a small PR that add a note in the installation page 🙂 Sorry for not doing this sooner!
