### 2024-12-06 #7400 @ivastar
@TomWagg a handful of editorial comments. Once you have completed them, please proceed to the checklist below.

Title: Please capitalize properly
Line 15: their - > its
Line 18: analytic - > analytical
Line 19: to -> into
Line 29: comma after "functions"
Line 23: binaries -> bianry? as in "binary and multiple" go together as adjectives to star systems
Line 25: lifetimes --> lives?
Line 31: across -> over
Line 32: Both Ivanova citations have the first name of the author while none of the others do, please edit the bib file for them
Line 37: systems with -> systems at
Line 41: separation prior to supernova -> pre-supernova separation
Line 52: position -> positions
Line 54-55: Other work has --> Other works have
Lune 57: whilst -> while
Line 60: natal -> birth. I've not seen "natal" as a science terminology but maybe I'm mistaken for this specific field
Lines 59 and 62: consecutive sentences starting with "Additionally", remove one
Line 64-65: "the spatial distribution of Milky Way black holes and neutron stars" -> the spatial distribution of black holes and neutron stars in the Milky Way
Line 74: typo, fulfuls -> fulfills
Line 77: gravitational wave does not need to be hyphenated
