### 2024-12-06 #7400 @ivastar
@TomWagg also, please add the AAS DOI by hand in the Markdown

https://github.com/openjournals/joss-reviews/issues/7400#issuecomment-2438687323
