+++ paper/JOSS/paper.md
@@ -32,8 +32,8 @@ bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: #10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-aas-journal: #Astrophysical Journal <- The name of the AAS journal.
+aas-doi: 10.3847/1538-4365/ad8b1f
+aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
