+++ paper/JOSS/paper.bib
@@ -409,7 +409,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Ivanova+2013,
-       author = {{Ivanova}, N. and {Justham}, S. and {Chen}, X. and {De Marco}, O. and {Fryer}, C.~L. and {Gaburov}, E. and {Ge}, H. and {Glebbeek}, E. and {Han}, Z. and {Li}, X. -D. and {Lu}, G. and {Marsh}, T. and {Podsiadlowski}, P. and {Potter}, A. and {Soker}, N. and {Taam}, R. and {Tauris}, T.~M. and {van den Heuvel}, E.~P.~J. and {Webbink}, R.~F.},
+       author = {{Ivanova}, Natalia and {Justham}, S. and {Chen}, X. and {De Marco}, O. and {Fryer}, C.~L. and {Gaburov}, E. and {Ge}, H. and {Glebbeek}, E. and {Han}, Z. and {Li}, X. -D. and {Lu}, G. and {Marsh}, T. and {Podsiadlowski}, P. and {Potter}, A. and {Soker}, N. and {Taam}, R. and {Tauris}, T.~M. and {van den Heuvel}, E.~P.~J. and {Webbink}, R.~F.},
         title = "{Common envelope evolution: where we stand and how we can move forward}",
       journal = {\aapr},
      keywords = {Close binaries, Stellar structure, interiors, evolution, Hydrodynamics, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
