+++ paper/JOSS/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'cogsworth: A Gala of COSMIC proportions combining binary stellar evolution and galactic dynamics'
+title: 'cogsworth: A Gala of COSMIC Proportions Combining Binary Stellar Evolution and Galactic Dynamics'
 tags:
   - Python
   - astronomy
@@ -38,19 +38,19 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 
 # Summary
 
-We present `cogsworth`, an open-source Python tool for producing self-consistent population synthesis and galactic dynamics simulations. With `cogsworth` one can (1) sample a population of binaries and star formation history, (2) perform rapid (binary) stellar evolution, (3) integrate orbits through the galaxy and (4) inspect the full evolutionary history of each star or compact object, as well as their positions and kinematics. We include the functionality for post-processing hydrodynamical zoom-in simulations as a basis for galactic potentials and star formation histories to better account for initial spatial stellar clustering and more complex potentials. Alternatively, several analytic models are available for both the potential and star formation history. `cogsworth` can transform the intrinsic simulated population to an observed population through the joint application of dust maps, bolometric correction functions and survey selection functions. This paper is jointly published in ApJS [@cogsworth_full].
+We present `cogsworth`, an open-source Python tool for producing self-consistent population synthesis and galactic dynamics simulations. With `cogsworth` one can (1) sample a population of binaries and star formation history, (2) perform rapid (binary) stellar evolution, (3) integrate orbits through the galaxy and (4) inspect the full evolutionary history of each star or compact object, as well as its positions and kinematics. We include the functionality for post-processing hydrodynamical zoom-in simulations as a basis for galactic potentials and star formation histories to better account for initial spatial stellar clustering and more complex potentials. Alternatively, several analytical models are available for both the potential and star formation history. `cogsworth` can transform the intrinsic simulated population into an observed population through the joint application of dust maps, bolometric correction functions, and survey selection functions. This paper is jointly published in ApJS [@cogsworth_full].
 
 # Statement of need
 
-The majority of stars are born in binaries and multiple star systems (e.g., @Duchene+2013; @Moe+2017; @Offner+2023), a large subset of which will exchange mass at some point in their lifetime (e.g., @Podsiadlowski+1992; @Sana+2012; @deMink+2014). These massive stars play a critical role in the formation and evolution of galaxies as a result of their feedback (e.g., @Dekel+1986; @Hopkins+2012; @Nomoto+2013; @Somerville+2015; @Naab+2017). However, binary evolution remains uncertain, with many parameters such as common-envelope efficiency, mass transfer efficiency, angular momentum loss due to mass transfer and the mean magnitude of supernova natal kicks unconstrained across several orders of magnitude (e.g., @Janka+2012; @Ivanova+2013; @Katsuda+2018; @Ivanova+2020; @Ropke+2023; @Marchant2023).
+The majority of stars are born in binaries and multiple star systems (e.g., @Duchene+2013; @Moe+2017; @Offner+2023), a large subset of which will exchange mass at some point in their lives (e.g., @Podsiadlowski+1992; @Sana+2012; @deMink+2014). These massive stars play a critical role in the formation and evolution of galaxies as a result of their feedback (e.g., @Dekel+1986; @Hopkins+2012; @Nomoto+2013; @Somerville+2015; @Naab+2017). However, binary evolution remains uncertain, with many parameters such as common-envelope efficiency, mass transfer efficiency, angular momentum loss due to mass transfer and the mean magnitude of supernova natal kicks unconstrained over several orders of magnitude (e.g., @Janka+2012; @Ivanova+2013; @Katsuda+2018; @Ivanova+2020; @Ropke+2023; @Marchant2023).
 
[…]
