+++ paper/JOSS/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
 aas-doi: 10.3847/1538-4365/ad8b1f
-aas-journal: Astrophysical Journal
+aas-journal: Astrophysical Journal Supplement Series
 ---
 
 # Summary
