### 2024-12-06 #7400 @TomWagg
Thanks @ivastar, I've implemented your comments in the paper draft (though I kept "natal", that's the usual terminology for supernova kicks 🙂). I've also checked the authors and created a new release, which you can see [here on GitHub](https://github.com/TomWagg/cogsworth/releases/tag/v2.0.1) and [here on Zenodo](https://zenodo.org/records/14289112).

Version: 2.0.1
DOI: https://doi.org/10.5281/zenodo.14289112
