### 2024-12-09 #7400 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
