### 2024-12-09 #7400 @warrickball
@TomWagg — I've started working through our post-acceptance items and aim to make a PR with editorial tweaks in the next few days.

@nicolagaspari, @smandhai — As a matter of policy, note that this publication is part of a collaboration with AAS Journals, which means that AAS Journals make a small donation to the running costs of JOSS. You can find out more in the [original announcement](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing).
