### 2024-12-10 #7400 @warrickball
I've made a first (and hopefully only!) [PR](https://github.com/TomWagg/cogsworth/pull/149) with editorial tweaks to the paper. Once this is merged I'll test the publication pipeline again.
