+++ cogsworth/pop.py
@@ -22,7 +22,7 @@ import gala.dynamics as gd
 from gala.potential.potential.io import to_dict as potential_to_dict, from_dict as potential_from_dict
 
 from cogsworth import sfh
-from cogsworth.kicks import integrate_orbit_with_events
+from cogsworth.kicks import integrate_orbit_with_events, integrate_orbit
 from cogsworth.events import identify_events
 from cogsworth.classify import determine_final_classes
 from cogsworth.observables import get_photometry
@@ -1917,4 +1917,144 @@ class GalacticTidePopulation(Population):
         super().__init__(**pop_kwargs)
         self.galactic_tides_timestep = galactic_tides_timestep
 
+    def perform_stellar_evolution(self):
+        raise NotImplementedError("Independent stellar evolution not supported for GalacticTidesPopulations")
+
+    def perform_galactic_evolution(self, progress_bar=True):
+        """Use :py:mod:`gala` to perform the orbital integration for each binary
+
+        Parameters
+        ----------
+        quiet : `bool`, optional
+            Whether to silence any warnings about failing orbits, by default False
+        """
+        # delete any cached variables that are based on orbits
+        self._final_pos = None
+        self._final_vel = None
+        self._observables = None
+
+        if self._initial_galaxy is None:            # pragma: no cover
+            logging.getLogger("cogsworth").warning(("cogsworth warning: Initial galaxy not yet sampled, "
+                                                    "performing sampling now."))
+            self.sample_initial_galaxy()
+
+        if self._initC is None and self._initial_binaries is None:          # pragma: no cover
+            logging.getLogger("cogsworth").warning(("cogsworth warning: Initial binaries not yet sampled, "
+                                                    "performing sampling now."))
+            self.sample_initial_binaries()
+
+        v_phi = (self.initial_galaxy.v_T / self.initial_galaxy.rho)
+        v_X = (self.initial_galaxy.v_R * np.cos(self.initial_galaxy.phi)
+               - self.initial_galaxy.rho * np.sin(self.initial_galaxy.phi) * v_phi)
+        v_Y = (self.initial_galaxy.v_R * np.sin(self.initial_galaxy.phi)
+               + self.initial_galaxy.rho * np.cos(self.initial_galaxy.phi) * v_phi)
+
+        # combine the representation and differentials into a Gala PhaseSpacePosition
+        w0s = gd.PhaseSpacePosition(pos=[a.to(u.kpc).value for a in [self.initial_galaxy.x,
+                                                                     self.initial_galaxy.y,
+                                                                     self.initial_galaxy.z]] * u.kpc,
+                                    vel=[a.to(u.km/u.s).value for a in [v_X, v_Y,
+                                                                        self.initial_galaxy.v_z]] * u.km/u.s)
+
+        # if we want to use multiprocessing
+        if self.pool is not None or self.processes > 1:
+            # track whether a pool already existed
+            pool_existed = self.pool is not None
+
+            # if not, create one
+            if not pool_existed:
+                self.pool = Pool(self.processes)
+
+            # setup arguments to combine primary and secondaries into a single list
+            args = [(w0s[i], self.max_ev_time - self.initial_galaxy.tau[i], self.max_ev_time,
[…]
