### 2024-12-13 #7400 @warrickball
While we wait for the AAS Journals DOI to start resolving, I opened another [small PR](https://github.com/TomWagg/cogsworth/pull/150) to update the reference to Gaspari et al. (2024), which I think has just been published/indexed in NASA ADS.

I'll check the AAS Journals DOI every few days but feel free to tag me here if you notice it start working.
