### 2025-01-02 #7400 @augustfly
@warrickball -- I've told our production staff to just go ahead as they are ready to publish the AAS Journal article. BTW this is a game of chicken and egg: in the past typically JOSS publishes faster and thus first though we are happy to publish first if you now require it.
