### 2025-01-02 #7400 @warrickball
Hi @augustfly — I don't have a strong feeling either way but I was under the impression that AAS publishes first. Specifically, [our docs say](https://joss.readthedocs.io/en/latest/editing.html#aas-publishing):
> ...
> * Pause the review (by applying the paused label) to await notification that the AAS paper is published.
>
> **Once the AAS paper is published**
> * Ask the EiC on rotation to publish the paper as normal (by tagging `@openjournals/joss-eics`).

If this is mistaken or misleading, let me know and I'll update it.
