### 2025-01-02 #7400 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
